Delete unneeded files - the coverage & .idea folders & gmaps js file
The coverage folder shouldn't live in the git repo because there is
no gaureentee it is in sync with the code in the repo (unless you
run the test coverage before pushing every time). Also in general
they clutter the repo and every PR that creates them.

The .idea folder is for RubyMine or other IntelliJ IDEs uses only.
Not everyone uses these IDEs and it changes for each persons computer
so it doesn't make sense to put this in the repository either.

In addition there was an extra gmaps js file in the `public/javascripts`
folder which isn't needed since we will be manually adding gmaps later.

Lastly, this commit updates the .gitignore file so that these files
continue to be ignored so they do not show up user's git status'.